Center filter on main page
Reason: Filter should be centered in desktop window.
Description: To get the filter centered, the filter-options div must be a fixed width. Its margin must be set to auto so that the browser will allocate equal space on either side. The filter-container now has a width of 100% and the background-color #3397DB so that the light blue band covers the entire width of the window.