Initial implementation of service worker
feat: Add service worker script and register in main.js

This initial implementation is not complete. The service worker fails to register in Chrome because the scope is shallower than the script file. The site still runs. It just doesn't use the service worker. Firefox also runs because it does not support service workers.