Service worker corrections
feat: Service worker corrections and favicon.ico addition

The service worker logic was not working in the initial release because the sw.js file was not in the root directory.

Another error found was that google chrome looks for an favicon.ico file to display in the tab at the top of the screen. Created a generic png and referenced it in the index.html file.