Add lang attribute. Use h2 for sub headers.
feat: Add the lang attribute to both pages and use h2 for sub headers

The lang attribute is now set to "en". This was a suggested change by the Udactiy reviewer. It helps screen readers determine what language and pronunciation to use when reading a page. It also helps search engines find language specific content.

The main page has a header marked as h1. All subsequent headers should be marked h2. This was another change that the Udacity reviewer suggested.