Add descriptive alt tags to all images
feat: Add descriptive alt tags to all images

I had already included alt tags on the main page, but the reviewer suggested a better string stating the restaurant name and in what neighborhood it was located. I changed the alt tags on the main page to match the suggestion.

The restaurant_info.js file did not provide any alt tag for the image. I added the same alt tag as the main page.