Add aria roles and labels
feat: Add aria roles and labels as appropriate

Based on the review comments, aria roles and labels were added to the maps on both pages. The breadcumb was wrapped in a nav element with an aria role and label. The Home href was also given an aria label. The filter option dropdownlist boxes where also given aria-label attributes.